refactor(completion): split engine into modules
